backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
alembic==1.13.0
google-auth==2.26.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-jose==3.3.0
aiosmtplib==3.0.1
